tests/wpt/web-platform-tests/service-workers/service-worker/resource-timing.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
function resourceUrl(path) {
    return get_host_info()['HTTPS_ORIGIN'] + base_path() + path;
}

function crossOriginUrl(path) {
    return get_host_info()['HTTPS_REMOTE_ORIGIN'] + base_path() + path;
}

function verify(performance, resource, mode, description) {
    var url = mode === 'cross-origin' ? crossOriginUrl(resource)
                                      : resourceUrl(resource);
    var entryList = performance.getEntries();
    var entry = performance.getEntriesByName(url)[0];
    assert_greater_than(entry.workerStart, 0, description);
    assert_greater_than_equal(entry.workerStart, entry.startTime, description);
    assert_less_than_equal(entry.workerStart, entry.fetchStart, description);
    if (mode === 'cross-origin') {
      assert_equals(entry.responseStart, 0, description);
      assert_greater_than_equal(entry.responseEnd, entry.fetchStart, description);
    } else {
      assert_greater_than_equal(entry.responseStart, entry.fetchStart, description);
      assert_greater_than_equal(entry.responseEnd, entry.responseStart, description);
    }
    assert_greater_than(entry.responseEnd, entry.fetchStart, description);
    assert_greater_than(entry.duration, 0, description);
    if (resource.indexOf('redirect.py') != -1) {
        assert_less_than_equal(entry.workerStart, entry.redirectStart,
                               description);
    } else {
        assert_equals(entry.redirectStart, 0, description);
    }
}

async_test(function(t) {
    var worker_url = 'resources/resource-timing-worker.js';
    var scope = 'resources/resource-timing-iframe.sub.html';
    var registration;

    service_worker_unregister_and_register(t, worker_url, scope)
      .then(function(r) {
          registration = r;
          return wait_for_state(t, r.installing, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          var performance = frame.contentWindow.performance;
          verify(performance, 'resources/dummy.js', 'same-origin',
                 'Generated response');
          verify(performance, 'resources/empty.js', 'same-origin',
                 'Network fallback');
          verify(performance, 'resources/redirect.py?Redirect=empty.js',
                 'same-origin', 'Redirect');
          verify(performance, 'resources/missing.jpg', 'same-origin',
                 'Network fallback image');

          // Test that worker start is available on cross-origin no-cors
          // subresources.
          verify(performance, 'resources/missing.jpg', 'cross-origin',
                 'Network fallback cross-origin image');

          frame.remove();
          return registration.unregister();
        })
      .then(function() {
          t.done();
        })
      .catch(unreached_rejection(t));
}, 'Controlled resource loads');

test(function() {
    var url = resourceUrl('resources/test-helpers.sub.js');
    var entry = window.performance.getEntriesByName(url)[0];
    assert_equals(entry.workerStart, 0, 'Non-controlled');
}, 'Non-controlled resource loads');

</script>
